refactor:테스트코드에서 분류되지 않았던 constants 정리
